move repository unit test in 'unit' repository,add category and thematic integration test, add conftest.py file to write configuration for pytest, update schemas.py, requirement.txt and readme.md

+++ app/categories/service.py
@@ -1,12 +1,15 @@
+import json
 from datetime import datetime
 from flask import abort, jsonify
-from typing import List, Dict
+from typing import List, Dict, Mapping
 from injector import inject
 from sqlalchemy.exc import SQLAlchemyError
+from flask_jwt_extended import  get_jwt_identity
 
 from app.categories.repository import CategoryRepository
 from app.houses.models import HouseModel
 from app.categories.models import CategoryModel
+from app.users.repository import UserRepository
 
 
 class CategoryService:
@@ -14,50 +17,62 @@ class CategoryService:
     @inject
     def __init__(self):
          self.category_repository = CategoryRepository()
+         self.user_repository = UserRepository()
 
 
-    def get_all_categories(self)-> List[CategoryModel]:
+    def get_all_categories(self):
         """
         Return all categories
         :return: a list of Category objects
         """
+        
         return self.category_repository.get_all()
 
     def get_category_by_id(self, category_id: int) -> CategoryModel:
-        # return self.category_repository.get_category_by_id(category_id)
-        return self.category_repository.get_category_by_id(category_id)
+        curent_user = self.user_repository.get_user_by_id(get_jwt_identity())
+        if curent_user.is_admin:
+           return self.category_repository.get_category_by_id(category_id)
+        return{"message": "Access Denied"}, 403
 
 
     def create_category(self, category):
-        try:
-            # print(category)
-            self.category_repository.save(category)
-            self.category_repository.commit()
-            
-            return{"message": "category created successfully."}, 201
-        except SQLAlchemyError:
-            abort(500,"An error occurred while inserting the category")
+        curent_user = self.user_repository.get_user_by_id(get_jwt_identity())
+        if curent_user.is_admin:
+            try:
+                self.category_repository.save(category)
+                self.category_repository.commit()
+                return category, 201
+            except SQLAlchemyError:
+                abort(500,"An error occurred while inserting the category")
+        return{"message": "Access Denied"}, 403
 
     def update_category(self, category_id:int, category_data:Dict[str, None]):
-        try:
-            category = self.category_repository.get_category_by_id(category_id)
-            category.show = category_data.get("show", 0)
-            category.libelle = category_data.get("libelle","Not define")
-            self.category_repository.save(category)
-            self.category_repository.commit()
-            return category
-        except:
-            abort(404, f"A category with id:{category_id} doesn't exist")
+        curent_user = self.user_repository.get_user_by_id(get_jwt_identity())
+        
+        if curent_user.is_admin:
+            try:
+                category = self.category_repository.get_category_by_id(category_id)
+                category.show = category_data.get("show", 0)
+                category.libelle = category_data.get("libelle","Not define")
+                self.category_repository.save(category)
+                self.category_repository.commit()
+                return category
+            except:
+                abort(404, f"A category with id:{category_id} doesn't exist")
+        return{"message": "Access Denied"}, 403
 
 
     def delete_category(self, category_id):
-        try:
-            category = self.category_repository.get_category_by_id(category_id)
-            self.category_repository.delete(category)
-            self.category_repository.commit()
-            return  {"message":"category deleted"}, 200
-        except:
-            abort(404, f"A category with id:{category_id} doesn't exist")
+        curent_user = self.user_repository.get_user_by_id(get_jwt_identity())
+        if curent_user.is_admin:
+            try:
+                category = self.category_repository.get_category_by_id(category_id)
+                self.category_repository.delete(category)
+                self.category_repository.commit()
+                return  {"message":"category deleted"}, 204
+            except:
+                abort(404, f"A category with id:{category_id} doesn't exist")
+        return{"message": "Access Denied"}, 403
 
     def get_houses_in_category(self, category_id:int) -> List[HouseModel]:
         """Get Houses which belong to category
